Solution - car wash station update: potential division by zero error resolve

+++ app/main.py
@@ -14,7 +14,7 @@ class Car:
 class CarWashStation:
     def __init__(
             self,
-            distance_from_city_center: int,
+            distance_from_city_center: float,
             clean_power: int,
             average_rating: float,
             count_of_ratings: int
@@ -39,7 +39,7 @@ class CarWashStation:
 
     def calculate_washing_price(self, car: Car) -> float:
         price = (car.comfort_class * (self.clean_power - car.clean_mark)
-                 * self.average_rating / self.distance_from_city_center)
+                 * self.average_rating / max(self.distance_from_city_center, 1))
 
         return round(price, 1)
